Organized the model classes into new folders and name spaces on the client side.
