+++ FribergFastigheter.Client/AutoMapper/DtoToViewModelAutoMapperProfile.cs
@@ -1,9 +1,13 @@
 ﻿using AutoMapper;
-using FribergFastigheter.Client.Models;
+using FribergFastigheter.Client.Models.Broker;
+using FribergFastigheter.Client.Models.BrokerFirm;
+using FribergFastigheter.Client.Models.Housing;
+using FribergFastigheter.Client.Models.Image;
 using FribergFastigheter.Shared.Dto.Broker;
 using FribergFastigheter.Shared.Dto.BrokerFirm;
 using FribergFastigheter.Shared.Dto.Housing;
 using FribergFastigheter.Shared.Dto.Image;
+using FribergFastigheter.Shared.Dto.Statistics;
 
 namespace FribergFastigheter.Client.AutoMapper
 {
@@ -24,8 +28,10 @@ namespace FribergFastigheter.Client.AutoMapper
 			CreateMap<MunicipalityDto, MunicipalityViewModel>();
 			CreateMap<ImageDto, ImageViewModel>();
 			CreateMap<PaginationDto, PaginationViewModel>();
+            CreateMap<BrokerFirmStatisticsDto, BrokerFirmStatisticsViewModel>();
+            CreateMap<StatisticItemDto, StatisticItemViewModel>();
 
-			CreateMap<BrokerFirmDto, BrokerFirmViewModel>()
+            CreateMap<BrokerFirmDto, BrokerFirmViewModel>()
 				.ForMember(dest => dest.Logotype, opt => opt.PreCondition(x => x.Logotype != null));
 
 			CreateMap<BrokerFirmSummaryDto, BrokerFirmSummaryViewModel>()
