+++ FribergFastigheter.Client/Components/ConfirmButton.razor
@@ -7,7 +7,7 @@
     string labelId = Guid.NewGuid().ToString();
 }
 <button type="button" class="btn btn-danger" data-bs-toggle="modal" data-bs-target="#@dialogId">
-    Radera
+    @ButtonText
 </button>
 <div class="modal fade" id="@dialogId" tabindex="-1" aria-labelledby="@labelId" aria-hidden="true">
     <div class="modal-dialog">
