+++ FribergFastigheter.Client/Layout/BrokerFirmMember/BrokerBackOfficeLayout.razor
@@ -1,0 +1,13 @@
+﻿@inherits LayoutComponentBase
+@*
+<!-- Author: Jimmie -->
+<!-- Co Authors: -->
+*@
+<div class="page">
+    <main>
+        <BrokerBackOfficeNavMenu />
+        <article class="content px-4">
+            @Body
+        </article>
+    </main>
+</div>
