+++ FribergFastigheter.Client/Layout/BrokerFirmMember/BrokerBackOfficeNavMenu.razor
@@ -1,0 +1,34 @@
+﻿@using FribergFastigheter.Client.Components
+@using FribergFastigheter.Shared.Constants
+@*
+<!-- Author: Jimmie -->
+<!-- Co Authors: -->
+*@
+<AuthorizeView>
+    <Authorized>
+        <nav class="navbar navbar-expand-lg bg-light">
+            <div class="container-fluid">
+                <a class="navbar-brand" href="/">Fribergs Fastigheter</a>
+                <button class="navbar-toggler" type="button" data-bs-toggle="collapse" data-bs-target="#mainNavigationMenu" 
+                    aria-controls="mainNavigationMenu" aria-expanded="false" aria-label="Toggle navigation">
+                    <span class="navbar-toggler-icon"></span>
+                </button>
+                <div class="collapse navbar-collapse" id="mainNavigationMenu">
+                    <ul class="navbar-nav justify-content-between w-100">
+                        <div class="navbar-nav">
+                            <li class="nav-item">
+                                <a class="nav-link" href="brokermember/housing/list">Bostäder</a>
+                            </li>
+                            <li class="nav-item">
+                                <a class="nav-link" href="brokermember/broker/list">Mäklare</a>
+                            </li>                                       
+                        </div>
+                        <div class="navbar-nav">
+                            <LoginButton />
+                        </div>
+                    </ul>
+                </div>
+            </div>
+        </nav>
+    </Authorized>
+</AuthorizeView>
