+++ FribergFastigheter.Client/Layout/NavMenu.razor
@@ -4,8 +4,6 @@
 <!-- Author: Jimmie -->
 <!-- Co Authors: -->
 *@
-
-
 <nav class="navbar navbar-expand-lg bg-light">
     <div class="container-fluid">
         <a class="navbar-brand" href="/">Fribergs Fastigheter</a>
@@ -16,26 +14,13 @@
         <div class="collapse navbar-collapse" id="mainNavigationMenu">
             <ul class="navbar-nav justify-content-between w-100">
                 <div class="navbar-nav">
-                    <AuthorizeView Policy="@ApplicationPolicies.BrokerAdmin">
-                        <Authorized>
-                            <li class="nav-item">
-                                <a class="nav-link" href="brokermember">Hem</a>
-                            </li>
-                            <li class="nav-item">
-                                <a class="nav-link" href="brokermember/housing/list">Bostäder</a>
-                            </li>
-                            <li class="nav-item">
-                                <a class="nav-link" href="brokermember/broker/list">Mäklare</a>
-                            </li>
-                        </Authorized>
-                        <NotAuthorized>
-                            <li class="nav-item">
-                                <a class="nav-link active" aria-current="page" href="/">Hem</a>
-                            </li>                           
-                        </NotAuthorized>
-                    </AuthorizeView>                    
+                    <li class="nav-item">
+                        <a class="nav-link" href="/">Hem</a>
+                    </li>
                 </div>
-                <LoginButton />
+                <div class="navbar-nav">
+                    <LoginButton />
+                </div>
             </ul>
         </div>
     </div>
