+++ FribergFastigheter.Client/Pages/BrokerFirmMember/Broker/BrokerDetailsPage.razor
@@ -1,7 +1,9 @@
 ﻿@page "/brokermember/broker/{id:int}"
+@layout BrokerBackOfficeLayout
 @using FribergFastigheter.Client.Components
 @using FribergFastigheter.Client.Components.Broker
 @using FribergFastigheter.Client.Components.Housing
+@using FribergFastigheter.Client.Layout.BrokerFirmMember
 @using Microsoft.AspNetCore.Authorization
 @*
 <!-- Author: Marcus -->
