+++ FribergFastigheter.Client/Pages/BrokerFirmMember/Broker/BrokerDetailsPage.razor.cs
@@ -1,5 +1,6 @@
 ﻿using AutoMapper;
-using FribergFastigheter.Client.Models;
+using FribergFastigheter.Client.Models.Broker;
+using FribergFastigheter.Client.Models.Housing;
 using FribergFastigheter.Client.Services.FribergFastigheterApi;
 using FribergFastigheter.Shared.Dto.Broker;
 using FribergFastigheter.Shared.Dto.Housing;
