+++ FribergFastigheter.Client/Pages/BrokerFirmMember/Broker/ListBrokerPage.razor
@@ -1,6 +1,8 @@
 ﻿@page "/brokermember/broker/list"
+@layout BrokerBackOfficeLayout
 @using FribergFastigheter.Client.Components
 @using FribergFastigheter.Client.Components.Broker
+@using FribergFastigheter.Client.Layout.BrokerFirmMember
 @using Microsoft.AspNetCore.Authorization
 @*
 <!-- Author: Marcus -->
