+++ FribergFastigheter.Client/Pages/BrokerFirmMember/BrokerFirmHomePage.razor
@@ -1,5 +1,7 @@
 ﻿@page "/brokermember"
+@layout BrokerBackOfficeLayout
 @using FribergFastigheter.Client.Components.Statistics
+@using FribergFastigheter.Client.Layout.BrokerFirmMember
 @using Microsoft.AspNetCore.Authorization
 @* 
 <!-- Author: Jimmie -->
