+++ FribergFastigheter.Client/Pages/BrokerFirmMember/Firm/BrokerFirmDetailsPage.razor
@@ -1,6 +1,8 @@
 ﻿@page "/brokermember/brokerfirm"
+@layout BrokerBackOfficeLayout
 @using FribergFastigheter.Client.Components
 @using FribergFastigheter.Client.Components.Housing
+@using FribergFastigheter.Client.Layout.BrokerFirmMember
 @using Microsoft.AspNetCore.Authorization
 @using FribergFastigheter.Client.Components.BrokerFirm
 @using FribergFastigheter.Client.Components.Broker
