+++ FribergFastigheter.Client/Pages/BrokerFirmMember/Firm/BrokerFirmDetailsPage.razor.cs
@@ -1,5 +1,7 @@
 ﻿using AutoMapper;
-using FribergFastigheter.Client.Models;
+using FribergFastigheter.Client.Models.Broker;
+using FribergFastigheter.Client.Models.BrokerFirm;
+using FribergFastigheter.Client.Models.Housing;
 using FribergFastigheter.Client.Services.FribergFastigheterApi;
 using FribergFastigheter.Shared.Dto.BrokerFirm;
 using Microsoft.AspNetCore.Components;
