+++ FribergFastigheter.Client/Pages/BrokerFirmMember/Housing/ListHousingPage.razor
@@ -1,6 +1,8 @@
 ﻿@page "/brokermember/housing/list"
+@layout BrokerBackOfficeLayout
 @using FribergFastigheter.Client.Components
 @using FribergFastigheter.Client.Components.Housing
+@using FribergFastigheter.Client.Layout.BrokerFirmMember
 @using Microsoft.AspNetCore.Authorization
 @*
     <!-- Author: Jimmie -->
