+++ FribergFastigheter.Client/Pages/BrokerFirmMember/Housing/ListHousingPage.razor.cs
@@ -1,6 +1,6 @@
 ﻿using AutoMapper;
 using FribergFastigheter.Client.Components;
-using FribergFastigheter.Client.Models;
+using FribergFastigheter.Client.Models.Housing;
 using FribergFastigheter.Client.Services.FribergFastigheterApi;
 using FribergFastigheter.Shared.Dto;
 using Microsoft.AspNetCore.Components;
