+++ FribergFastigheter.Client/Pages/Login.razor.cs
@@ -1,5 +1,5 @@
 ﻿using AutoMapper;
-using FribergFastigheter.Client.Models;
+using FribergFastigheter.Client.Models.BrokerFirm;
 using FribergFastigheter.Client.Services.FribergFastigheterApi;
 using FribergFastigheter.Shared.Dto.Login;
 using Microsoft.AspNetCore.Components;
